Finish README
I added a third line to the editor. The guide told me that in this box I should write the changes I did so I can follow the steps later. (Looking at the image, I understood that I needed to change the box "Uptade README.md" to "Finish README" that)